% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentGrowthCurveWrapper.R
\name{LatentGrowthCurve}
\alias{LatentGrowthCurve}
\title{Latent Growth}
\usage{
LatentGrowthCurve(
  data = NULL,
  version = "1",
  additionalFitMeasures = FALSE,
  bootstrapCiType = "percentileBiasCorrected",
  bootstrapSamples = 1000,
  categorical = list(types = list(), value = list()),
  ciLevel = 0.95,
  colorPalette = "colorblind",
  covariates = list(types = list(), value = list()),
  covaryingLatentCurve = TRUE,
  cubic = FALSE,
  curvePlot = FALSE,
  curvePlotCategorical = list(types = "unknown", value = ""),
  curvePlotMaxLines = 150,
  dependentCorrelation = TRUE,
  emulation = "lavaan",
  errorCalculationMethod = "standard",
  estimator = "default",
  exogenousLatentCorrelation = TRUE,
  group = list(types = "unknown", value = ""),
  impliedCovariance = FALSE,
  intercept = TRUE,
  latentInterceptFixedToZero = TRUE,
  linear = TRUE,
  manifestInterceptFixedToZero = FALSE,
  misfitPlot = FALSE,
  naAction = "fiml",
  pathPlot = FALSE,
  pathPlotMean = FALSE,
  pathPlotParameter = FALSE,
  plotHeight = 320,
  plotWidth = 480,
  quadratic = FALSE,
  rSquared = FALSE,
  regressions = list(types = list(), value = list()),
  residualCovariance = FALSE,
  residualSingleIndicatorOmitted = TRUE,
  residualVariance = TRUE,
  scalingParameter = TRUE,
  standardizedEstimate = FALSE,
  standardizedEstimateType = "all",
  syntax = FALSE,
  threshold = TRUE,
  timings = list(optionKey = "variable", types = list(), value = list()),
  variables = list(types = list(), value = list())
)
}
\arguments{
\item{additionalFitMeasures, }{Display additional model fit measures
Defaults to \code{FALSE}.}

\item{bootstrapCiType, }{Select the type of bootstrap confidence interval to compute}

\item{bootstrapSamples, }{Specify the number of bootstrap samples}

\item{categorical, }{Categorical variables acting as factors}

\item{ciLevel, }{Set the confidence level for the interval estimates}

\item{colorPalette, }{Select a color palette for the plots}

\item{covaryingLatentCurve, }{Include covariance between latent growth components
Defaults to \code{TRUE}.}

\item{cubic, }{Include a cubic component in the model
Defaults to \code{FALSE}.}

\item{curvePlot, }{Generate a curve plot of the growth trajectories
Defaults to \code{FALSE}.}

\item{curvePlotCategorical, }{Select a categorical variable to color the curves}

\item{curvePlotMaxLines, }{Set the maximum number of lines to display in the curve plot}

\item{emulation, }{Select the software to emulate estimation behavior
\itemize{
  \item \code{"lavaan"}: Use Lavaan default estimation method
  \item \code{"mplus"}: Emulate Mplus estimation methods
  \item \code{"eqs"}: Emulate EQS estimation methods
}}

\item{errorCalculationMethod, }{Select the method for calculating standard errors and confidence intervals
\itemize{
  \item \code{"standard"}: Use standard maximum likelihood estimation for standard errors
  \item \code{"robust"}: Use robust estimation for standard errors
  \item \code{"bootstrap"}: Use bootstrap method for estimating standard errors and confidence intervals
}}

\item{estimator, }{Choose the estimator for model fitting
\itemize{
  \item \code{"default"}: Use the default estimator based on the model and data
  \item \code{"ml"}: Maximum Likelihood estimator
  \item \code{"gls"}: Generalized Least Squares estimator
  \item \code{"wls"}: Weighted Least Squares estimator
  \item \code{"uls"}: Unweighted Least Squares estimator
  \item \code{"dwls"}: Diagonally Weighted Least Squares estimator
}}

\item{impliedCovariance, }{Display the implied covariance matrix
Defaults to \code{FALSE}.}

\item{intercept, }{Include an intercept in the model
Defaults to \code{TRUE}.}

\item{linear, }{Include a linear component in the model
Defaults to \code{TRUE}.}

\item{naAction, }{Select the method for handling missing data in estimation
\itemize{
  \item \code{"fiml"}: Use Full Information Maximum Likelihood to handle missing data
  \item \code{"listwise"}: Exclude cases with missing data (listwise deletion)
}}

\item{pathPlot, }{Generate a path diagram of the model
Defaults to \code{FALSE}.}

\item{pathPlotMean, }{Display means in the model plot
Defaults to \code{FALSE}.}

\item{pathPlotParameter, }{Display parameter estimates in the model plot
Defaults to \code{FALSE}.}

\item{quadratic, }{Include a quadratic component in the model
Defaults to \code{FALSE}.}

\item{rSquared, }{Display R-squared values for variables
Defaults to \code{FALSE}.}

\item{regressions, }{Variables to use as regressors in the model}

\item{residualCovariance, }{Display the residual covariance matrix
Defaults to \code{FALSE}.}

\item{standardizedEstimate, }{Include standardized estimates in the output
Defaults to \code{FALSE}.}

\item{standardizedEstimateType, }{Select the type of standardized estimates to include
\itemize{
  \item \code{"all"}: Include all standardized estimates
  \item \code{"latents"}: Include standardized estimates for latent variables only
  \item \code{"nox"}: Exclude standardized estimates for exogenous covariates
}}

\item{syntax, }{Show the lavaan model syntax
Defaults to \code{FALSE}.}

\item{timings, }{Variable timing}

\item{variables, }{Variables to include in the latent growth curve model}
}
\description{
Latent Growth
}
